Fix collection health overflow and query stats DOP type mismatch
- collection_health view: cast duration_ms and rows_collected to bigint
  before AVG/SUM to prevent arithmetic overflow after 7 days of data
- collection_health C# reader: use Convert.ToInt64 for avg_duration_ms,
  update model property from int to long
- query_stats reader: min_dop/max_dop columns are bigint in the table
  but reader called GetInt16() — changed to Convert.ToInt16(GetValue())

+++ Dashboard/Models/CollectionHealthItem.cs
@@ -19,7 +19,7 @@ namespace PerformanceMonitorDashboard.Models
         public decimal FailureRatePercent { get; set; }
         public long TotalRuns7d { get; set; }
         public long FailedRuns7d { get; set; }
-        public int AvgDurationMs { get; set; }
+        public long AvgDurationMs { get; set; }
         public long TotalRowsCollected7d { get; set; }
     }
 }
